sauditourism/api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
httpx>=0.25.0
anthropic>=0.39.0
PyPDF2>=3.0.0
